finishing with java

+++ db_server_sql/honey/api/api.py
@@ -25,25 +25,28 @@ def get_tn(table_name):
         return Response(get_all(con, table_name),content_type="application/json")
 
     if request.method == "POST":
+        print("usando POST")
         new_row = insert_new_row(con,request,table_name)
         if new_row:
-            return redirect(f"http://localhost:5000/honey/{table_name}")
             return Response(get_all(con, table_name),content_type="application/json")
 
-@app.route("/honey/api/<table_name>/<id>")
+@app.route("/honey/api/<table_name>/<id>", methods=["GET","DELETE"])
 def get_id(table_name,id):
-    requested_id = get_by_id(con,table_name,id)
-    if requested_id:
-        return Response(requested_id, content_type="application/json")
-    return {"success":False}
+    if request.method == "GET":
+        requested_id = get_by_id(con,table_name,id)
+        if requested_id:
+            return Response(requested_id, content_type="application/json")
+        return {"success":False}
+    if request.method == "DELETE":
+        print("usando delete")
+        is_deleted = delete_id(con,table_name,id)
+        if is_deleted:
+            # return redirect(f"http://localhost:5000/honey/{table_name}")
+            return {"success":True}
+        return {"success":False}
 
-@app.route("/honey/api/<table_name>/<id>/delete")
-def delete_person(table_name,id):
-    is_deleted = delete_id(con,table_name,id)
-    if is_deleted:
-        return redirect(f"http://localhost:5000/honey/{table_name}")
-        return {"success":True}
-    return {"success":False}
+
+
 
 if __name__ == "__main__":
     app.run(debug=True,port=3000)
